Get game

+++ backend/src/main/java/xyz/advtopics/objects/DTOs/GameDTO.java
@@ -2,5 +2,5 @@ package xyz.advtopics.objects.DTOs;
 
 public class GameDTO {
     public long dateTime;
-    public int[] teamIds;
+    public long[] teamIds;
 }
